Move all pages/components to individual directories
