+++ flask-server/requirements.txt
@@ -31,3 +31,4 @@ typing_extensions==4.12.2
 virtualenv==20.27.0
 Werkzeug==3.0.4
 pyopenssl==24.2.1
+flask_migrate==4.0.7
