+++ front-end/src/App.js
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 
-import NavBar from './components/NavBar'
-import { pages } from './Constants'
+import { NavBar } from './components'
+import { pages } from './utilities'
 import './App.css'
