+++ front-end/src/components/JournalIcon/JournalIcon.css
