+++ front-end/src/components/JournalIcon/JournalIcon.js
