+++ front-end/src/components/index.js
@@ -1,0 +1,9 @@
+import EmotionDropdown from './EmotionDropdown/EmotionDropdown'
+import JournalIcon from './JournalIcon/JournalIcon'
+import NavBar from './NavBar/NavBar'
+
+export {
+    EmotionDropdown,
+    JournalIcon,
+    NavBar
+}
