+++ front-end/src/pages/AllJournals/AllJournals.css
