+++ front-end/src/pages/JournalView/JournalView.js
