+++ front-end/src/pages/JournalWrite/JournalWrite.css
