+++ front-end/src/pages/index.js
@@ -1,0 +1,17 @@
+import Home from './Home/Home'
+import LogIn from './LogIn/LogIn'
+import SignUp from './SignUp/SignUp'
+import JournalWrite from './JournalWrite/JournalWrite'
+import AllJournals from './AllJournals/AllJournals'
+import JournalView from './JournalView/JournalView'
+import Page404 from './Page404/Page404'
+
+export {
+    Home,
+    LogIn,
+    SignUp,
+    JournalWrite,
+    AllJournals,
+    JournalView,
+    Page404
+}
