+++ front-end/src/utilities/index.js
@@ -1,3 +1,4 @@
 import axiosClient from "./axiosClient"
+import { pages, emotions } from "./Constants"
 
-export {axiosClient}
+export {axiosClient, pages, emotions}
